requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
annotated-types==0.7.0
anyio==4.6.2.post1
click==8.1.7
colorama==0.4.6
fast-depends==2.4.12
fastapi==0.115.5
faststream==0.5.30
h11==0.14.0
idna==3.10
multidict==6.1.0
pamqp==3.3.0
propcache==0.2.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.0
websockets==14.1
yarl==1.17.2
